Rename shims::io to shims::fs
