+++ src/shims/mod.rs
@@ -3,7 +3,7 @@ pub mod env;
 pub mod foreign_items;
 pub mod intrinsics;
 pub mod tls;
-pub mod io;
+pub mod fs;
 
 use rustc::{mir, ty};
